fix GC-root bugs in import resolution and OP_CALL/TAIL_CALL
Three related heisenbugs where bare C locals / raw Val* tail pointers
dangled after proc_grow (realloc) moved the heap. gc_fixup_heap_pointers
updates gc_roots but not C locals, so the corruption only surfaced under
ASAN or specific heap layouts.

src/api.c — load_module() and resolve_imports_in_ast():
  The previous "fix" (commit b036b20) only rooted cur/form in
  vm_resolve_imports_fn (the TA-parser path). The C-reader path used by
  vm.load_source (driver.ta's compile_file_to_tabc) still held cur/form/
  result in bare locals and a raw Val* tail pointer. Rewrite both to root
  head/last-cell/cur/form in sp->gc_roots and re-read after every val_pair
  / load_module call. Use a captured base offset since load_module is
  reentrant (called from resolve_imports_in_ast and itself) and both
  callers share sp->gc_roots.

src/vm.c — OP_CALL / OP_TAIL_CALL:
  args[] and free_vals[] were popped off the stack into C locals and then
  re-pushed one at a time to build the new frame. Any of those pushes can
  trigger gc_collect/proc_grow, dangling the not-yet-pushed entries and
  the closure_val (whose clos->entry is read afterwards). Root all of them
  in gc_roots around the push loop, re-read closure_val after the pushes.

ta.h — proc_push/proc_pop/proc_peek:
  p->sp * sizeof(Val) did unsigned pointer arithmetic (int * size_t),
  flagged by UBSan as pointer-overflow. Cast to (int)sizeof(Val) to keep
  the offset signed, matching the collision checks above.

Non-ASAN: 207/207 tests pass, bootstrap self-hosting is stable
(compile driver.ta → use as bootstrap → recompile → bytecode-identical).
ASAN still exposes a separate, deeper hang in compile_code (see note).

+++ ta.h
@@ -467,17 +467,17 @@ static inline void proc_push(Proc *p, Val v) {
         }
         v = gc_root_pop(p);
     }
-    *(Val *)(p->mem + p->mem_size + p->sp * sizeof(Val)) = v;
+            *(Val *)(p->mem + p->mem_size + p->sp * (int)sizeof(Val)) = v;
 }
 
 static inline Val proc_pop(Proc *p) {
-    Val v = *(Val *)(p->mem + p->mem_size + p->sp * sizeof(Val));
+    Val v = *(Val *)(p->mem + p->mem_size + p->sp * (int)sizeof(Val));
     p->sp++;
     return v;
 }
 
 static inline Val proc_peek(Proc *p, int offset) {
-    return *(Val *)(p->mem + p->mem_size + (p->sp + offset) * sizeof(Val));
+    return *(Val *)(p->mem + p->mem_size + (p->sp + offset) * (int)sizeof(Val));
 }
 
 /* ============================================================
@@ -516,7 +516,7 @@ static inline int proc_grow(Proc *p) {
     }
     uint8_t *new_mem = realloc(p->mem, new_size);
     if (!new_mem) return -1;
-        /* Relocate stack data to the new high end of the memory block.
+    /* Relocate stack data to the new high end of the memory block.
      * The stack grows downward from mem+mem_size; after doubling
      * mem_size, the stack base moves but the data hasn't. */
     {
